requirements.txt
Flask==2.0.1
pandas==2.2.1
Pillow==9.0.0 